pyproject.toml
[tool.poetry]
name = "mondayapi"
version = "0.1.0"
description = ""
authors = ["Miguel Henriques <github@jmh.com.br>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.31.0"
python-dotenv = "^1.0.1"
pony = "^0.7.17"
streamlit = "^1.31.1"
pandas = "^2.2.1"
plotly = "^5.19.0"


[tool.poetry.group.dev.dependencies]
ipython = "^8.21.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
